Basic Template and Homepage
This commit includes the core template we are using for the design of the website.

The main changes made to this template are on the homepage.
I have reformatted the homepage to fit the structure of the tech club website.

No design or content has been added at this point.